requirements.txt
#デプロイ用に必要なライブラリを列挙

Flask
Flask-SQLAlchemy
Flask-Mail
Flask-Login
tensorflow
keras
numpy
pillow
scikit-learn
gunicorn
pandas
pyotp
